refactor(ramfs): split middleware boundaries
